update post-verification steps
