+++ lab-template/src/algorithms/mining/block.py
@@ -1,3 +1,4 @@
+
 from dataclasses import dataclass
 from hashlib import sha256
 import json
@@ -5,25 +6,20 @@ import time
 import asyncio
 from ipv8.messaging.payload_dataclass import overwrite_dataclass
 from merkle_tree import MerkleTree
-from collections import defaultdict
-from ipv8.community import Community, CommunitySettings
-from ipv8.lazy_community import lazy_wrapper
-from ipv8.types import Peer
-
-
-class bcolors:
-    ONTRANSACTION = "\033[95m"
-    OKSIGNATURE = "\033[92m"
-    BADSIGNATURE = "\033[91m"
-
-    ONBLOCKMESSAGE = "\033[94m"
-    OKBLOCK = "\033[92m"
-
-    WARNING = "\033[93m"
-    ERROR = "\033[91m"
 
 # Custom dataclass implementation
 dataclass = overwrite_dataclass(dataclass)
+
+@dataclass(msg_id=3)  # Add a unique msg_id here
+class BlockMessage:
+    """ Represents a block message. """
+    timestamp: int
+    difficulty: int
+    nonce: int
+    prev_hash: str
+    merkle_root: str
+    transaction_tx: str
+    block_hash: str
 
 @dataclass
 class Transaction:
@@ -41,30 +37,16 @@ class SignedTransaction:
     signature: str
     public_key: str
 
-@dataclass
-class BlockMessage:
-    """ Represents a block message. """
-    timestamp: int
-    difficulty: int
-    nonce: int
-    prev_hash: str
-    merkle_root: str
-    coinbase_tx: str
+class bcolors:
+    ONTRANSACTION = "\033[95m"
+    OKSIGNATURE = "\033[92m"
+    BADSIGNATURE = "\033[91m"
 
-class Block:
-    """ Represents a block of transactions. """
-    def __init__(self, timestamp, difficulty, nonce, prev_hash, merkle_root, coinbase_tx):
-        self.timestamp = timestamp
-        self.difficulty = '0' * difficulty
-        self.nonce = nonce
-        self.prev_hash = prev_hash
-        self.merkle_root = merkle_root
-        self.coinbase_tx = coinbase_tx
-        self.hash = self.calculate_hash()
-    
-    def calculate_hash(self):
-        block_string = f'{self.timestamp}{self.difficulty}{self.nonce}{self.prev_hash}{self.merkle_root}{self.coinbase_tx}'
-        return sha256(block_string.encode()).hexdigest()
+    ONBLOCKMESSAGE = "\033[94m"
+    OKBLOCK = "\033[92m"
+
+    WARNING = "\033[93m"
+    ERROR = "\033[91m"
 
 class Blockchain:
     """ Manages the blockchain and its operations. """
@@ -73,44 +55,72 @@ class Blockchain:
         self.difficulty_target = difficulty_target
         self.chain = [self.create_genesis_block()]
         self.active_mining = False
+        self.community = None  # Will be set by ValidatorCommunity
+
+
 
     def create_merkle_root(self, transactions):
         """ Create a Merkle root from a list of signed transactions. """
+        print(f"Transatitionne1: {transactions}")
+
         tree = MerkleTree()
         for signed_tx in transactions:
             tx_string = json.dumps(signed_tx.__dict__, default=lambda o: o.__dict__)
             tree.add_leaf(tx_string)
         return tree.get_root()
+    
+ 
+        
 
     def create_genesis_block(self):
         timestamp = int(time.time())
         difficulty = self.difficulty_target
         nonce = 0
         prev_hash = '0' * 64
+        
         coinbase_tx = SignedTransaction(
             transaction=Transaction(sender="boss", receiver="0", amount=50, nonce=1, ts=timestamp),
             signature="coinbase_signature",
             public_key="coinbase_public_key"
         )
+        
         transactions = [coinbase_tx]
+        
         merkle_root = self.create_merkle_root(transactions)
-        return Block(timestamp, difficulty, nonce, prev_hash, merkle_root, json.dumps(coinbase_tx.__dict__, default=lambda o: o.__dict__))
+        
+        genesis_block = BlockMessage(
+            timestamp = timestamp, 
+            difficulty = difficulty, 
+            nonce=75806, 
+            prev_hash = prev_hash, 
+            merkle_root = merkle_root, 
+            transaction_tx=json.dumps(coinbase_tx.__dict__, default=lambda o: o.__dict__),
+            block_hash="00006a347b0bd5ee6c51104c4ed936597371ad02a5f9be8db3ea41be1e963462"
+        )
+        
+        # append the genesis block to the chain
+        return genesis_block
+    
+    
 
     async def create_new_block(self, transactions):
         timestamp = int(time.time())
         nonce = 0
-        prev_hash = self.chain[-1].hash
+        prev_hash = self.chain[-1].block_hash
         merkle_root = self.create_merkle_root(transactions)
         
         # Convert transactions to a JSON serializable format
         transactions_json = [tx.__dict__ for tx in transactions]
-        
-        new_block = Block(timestamp, self.difficulty_target, nonce, prev_hash, merkle_root, json.dumps(transactions_json))
-        
+
+        new_block = BlockMessage(timestamp = timestamp, difficulty = self.difficulty_target, nonce=nonce, 
+            prev_hash = prev_hash, 
+            merkle_root = merkle_root, 
+            transaction_tx=json.dumps(transactions_json),
+            block_hash="new_block_hash"
+        )
+
         await self.mine_block(new_block)
         return new_block
-
-        
 
 
     async def add_block(self, transactions):
@@ -124,11 +134,7 @@ class Blockchain:
         return new_block
     
     def compute_hash(self, block, nonce):
-        block_string = f'{block.timestamp}{block.difficulty}{nonce}{block.prev_hash}{block.merkle_root}{block.coinbase_tx}'
-        return sha256(block_string.encode()).hexdigest()
-
-    def compute_hash(self, block, nonce):
-        block_string = f'{block.timestamp}{block.difficulty}{nonce}{block.prev_hash}{block.merkle_root}{block.coinbase_tx}'
+        block_string = f'{block.timestamp}{block.difficulty}{nonce}{block.prev_hash}{block.merkle_root}{block.transaction_tx}'
         return sha256(block_string.encode()).hexdigest()
 
     async def mine_block(self, block):
@@ -145,37 +151,22 @@ class Blockchain:
         while True:
             self.active_mining = True
             
-            
             # Every 2 seconds, print the time elapsed and the number of hashes computed
             if time.time() - start_time > 2:
                 print(f"Hashes computed: {hashes_computed}")
-                # print the time since start_time was defined and formatted in seconds
                 print(f"Time elapsed: {time.time() - start_time2:.2f} seconds")
-                
-                
                 start_time = time.time()
-                hashes_computed = 0
                 
-            
             hash_result = self.compute_hash(block, current_nonce)
             
             if hash_result.startswith(target):
                 block.nonce = current_nonce
-                block.hash = hash_result
-                print(f"Block mined by miner {self.node_id} with hash {block.hash}")
+                block.block_hash = hash_result
+                print(f"Block mined by miner {self.node_id} with hash {block.block_hash}")
                 
-                
-                # edit the block's nonce and hash
-                block.nonce = current_nonce
-                block.hash = hash_result
-                block.difficulty = self.difficulty_target
-                
-              
-                
-                # decode the block and print it
                 print(
                     f'{bcolors.OKBLOCK}------------------------------------\n'
-                    f"{bcolors.OKBLOCK}Block mined by miner {self.node_id} with hash {block.hash}\n"
+                    f"{bcolors.OKBLOCK}Block mined by miner {self.node_id} with hash {block.block_hash}\n"
                     f'''
                     {bcolors.OKBLOCK}New Block:\n
                     {bcolors.OKBLOCK}Timestamp: {block.timestamp}\n
@@ -183,50 +174,20 @@ class Blockchain:
                     {bcolors.OKBLOCK}Nonce: {block.nonce}\n
                     {bcolors.OKBLOCK}Previous Hash: {block.prev_hash}\n
                     {bcolors.OKBLOCK}Merkle Root: {block.merkle_root}\n
-                    {bcolors.OKBLOCK}Transaction: {block.coinbase_tx}\n
-                    {bcolors.OKBLOCK}Hash: {block.hash}\n'''
+                    {bcolors.OKBLOCK}Transaction Body: {block.transaction_tx}\n
+                    {bcolors.OKBLOCK}Hash: {block.block_hash}\n'''
                     f'{bcolors.OKBLOCK}------------------------------------\n'
-                )        
+                )      
                 
-
                 # Add the block to the chain
                 self.chain.append(block)
                 
-                # Print all the blocks
-                for blk in self.chain:
-                    print(blk)
+                # Broadcast the block to peers
+                await self.community.broadcast_block(block)
                 
                 self.active_mining = False
                 return
             
             current_nonce += 1
             hashes_computed += 1
-            await asyncio.sleep(0)  # Yield control to allow other tasks to run
-
-
-
-
-
-# Example usage:
-# blockchain = Blockchain()
-# transactions = [
-#     SignedTransaction(
-#         transaction=Transaction(sender="user1", receiver="user2", amount=20, nonce=1, ts=int(time.time())),
-#         signature="signature1",
-#         public_key="public_key1"
-#     ),
-#     SignedTransaction(
-#         transaction=Transaction(sender="user3", receiver="user4", amount=30, nonce=1, ts=int(time.time())),
-#         signature="signature2",
-#         public_key="public_key2"
-#     ),
-#     SignedTransaction(
-#         transaction=Transaction(sender="user5", receiver="user6", amount=40, nonce=1, ts=int(time.time())),
-#         signature="signature3",
-#         public_key="public_key3"
-#     )
-# ]
-
-# new_block = blockchain.add_block(transactions, 1)
-
-# print(f'New Block:\nTimestamp: {new_block.timestamp}\nDifficulty: {new_block.difficulty}\nNonce: {new_block.nonce}\nPrevious Hash: {new_block.prev_hash}\nMerkle Root: {new_block.merkle_root}\nCoinbase Transaction: {new_block.coinbase_tx}\nHash: {new_block.hash}')
+            await asyncio.sleep(0)
